+++ src/Annotation/Method.php
@@ -7,9 +7,9 @@ namespace RichId\AutoconfigureBundle\Annotation;
 use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
 
 /**
- * Class Property.
+ * Class Method.
  *
- * Bind a service or a parameter to the designated property.
+ * Bind a service or a parameter to the designated method.
  *
  * @author     Nicolas Guilloux <nicolas.guilloux@rich-id.fr>
  * @copyright  2014 - 2021 Rich ID (https://www.rich-id.fr)
@@ -18,15 +18,19 @@ use Doctrine\Common\Annotations\Annotation\NamedArgumentConstructor;
  * @NamedArgumentConstructor()
  */
 #[\Attribute(\Attribute::IS_REPEATABLE | \Attribute::TARGET_CLASS)]
-final class Property extends AbstractServiceInjectionAnnotation
+final class Method extends AbstractServiceInjectionAnnotation
 {
     /** @var string */
-    public $property;
+    public $method;
 
-    public function __construct(string $property, string $value, string $type = self::SERVICE_TYPE)
+    /** @var int */
+    public $position = 0;
+
+    public function __construct(string $method, string $value, string $type = self::SERVICE_TYPE, int $position = 0)
     {
         parent::__construct($value, $type);
 
-        $this->property = $property;
+        $this->method = $method;
+        $this->position = $position;
     }
 }