+++ src/Configurators/Partials/ArgumentAutoConfigurator.php
@@ -4,12 +4,9 @@ declare(strict_types=1);
 
 namespace RichId\AutoconfigureBundle\Configurators\Partials;
 
-use RichId\AutoconfigureBundle\Annotation\Argument;
-use RichId\AutoconfigureBundle\Configurators\Basics\ServiceAutoConfiguratorInterface;
 use RichId\AutoconfigureBundle\Model\ServiceConfiguration;
-use Symfony\Component\DependencyInjection\Container;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
-use Symfony\Component\DependencyInjection\Reference;
 
 /**
  * Class ArgumentAutoConfigurator.
@@ -17,10 +14,10 @@ use Symfony\Component\DependencyInjection\Reference;
  * @author     Nicolas Guilloux <nicolas.guilloux@rich-id.fr>
  * @copyright  2014 - 2021 Rich ID (https://www.rich-id.fr)
  */
-final class ArgumentAutoConfigurator implements ServiceAutoConfiguratorInterface
+final class ArgumentAutoConfigurator extends AbstractServiceInjectionAutoConfigurator
 {
     public function autoconfigure(
-        Container $container,
+        ContainerBuilder $container,
         Definition $definition,
         ServiceConfiguration $configuration
     ): void {
@@ -29,25 +26,8 @@ final class ArgumentAutoConfigurator implements ServiceAutoConfiguratorInterface
 
             $definition->setArgument(
                 $sanitizedArgument,
-                $this->getObject($container, $options)
+                $this->resolveObject($container, $options)
             );
-        }
-    }
-
-    public function getObject(Container $container, array $options)
-    {
-        $type = $options['type'] ?? null;
-        $value = $options['value'] ?? null;
-
-        switch ($type) {
-            case Argument::SERVICE_TYPE:
-                return new Reference($value);
-
-            case Argument::PARAMETER_TYPE:
-                return $container->getParameter($value);
-
-            default:
-                throw new \UnexpectedValueException('The type used a service configuration is wrong.');
         }
     }
 }