+++ src/Configurators/Partials/PropertyAutoConfigurator.php
@@ -4,12 +4,9 @@ declare(strict_types=1);
 
 namespace RichId\AutoconfigureBundle\Configurators\Partials;
 
-use RichId\AutoconfigureBundle\Annotation\Property;
-use RichId\AutoconfigureBundle\Configurators\Basics\ServiceAutoConfiguratorInterface;
 use RichId\AutoconfigureBundle\Model\ServiceConfiguration;
-use Symfony\Component\DependencyInjection\Container;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
-use Symfony\Component\DependencyInjection\Reference;
 
 /**
  * Class PropertyAutoConfigurator.
@@ -17,30 +14,17 @@ use Symfony\Component\DependencyInjection\Reference;
  * @author     Nicolas Guilloux <nicolas.guilloux@rich-id.fr>
  * @copyright  2014 - 2021 Rich ID (https://www.rich-id.fr)
  */
-final class PropertyAutoConfigurator implements ServiceAutoConfiguratorInterface
+final class PropertyAutoConfigurator extends AbstractServiceInjectionAutoConfigurator
 {
     public function autoconfigure(
-        Container $container,
+        ContainerBuilder $container,
         Definition $definition,
         ServiceConfiguration $configuration
     ): void {
         foreach ($configuration->getProperties() as $property => $options) {
-            $type = $options['type'] ?? null;
-            $value = $options['value'] ?? null;
             $sanitizedProperty = \str_replace('$', '', $property);
-
-            switch ($type) {
-                case Property::SERVICE_TYPE:
-                    $definition->setProperty($sanitizedProperty, new Reference($value));
-                    break;
-
-                case Property::PARAMETER_TYPE:
-                    $definition->setProperty($sanitizedProperty, $container->getParameter($value));
-                    break;
-
-                default:
-                    throw new \UnexpectedValueException('The property type used a service configuration is wrong.');
-            }
+            $value = $this->resolveObject($container, $options);
+            $definition->setProperty($sanitizedProperty, $value);
         }
     }
 }