+++ src/Factory/ServiceConfigurationFactory.php
@@ -8,6 +8,7 @@ use RichId\AutoconfigureBundle\Factory\Basics\ServiceConfigurationFactoryInterfa
 use RichId\AutoconfigureBundle\Factory\Partials\ArgumentAnnotationServiceConfigurationFactory;
 use RichId\AutoconfigureBundle\Factory\Partials\DecorationAnnotationServiceConfigurationFactory;
 use RichId\AutoconfigureBundle\Factory\Partials\EventListenerServiceConfigurationFactory;
+use RichId\AutoconfigureBundle\Factory\Partials\MethodAnnotationServiceConfigurationFactory;
 use RichId\AutoconfigureBundle\Factory\Partials\PropertyAnnotationServiceConfigurationFactory;
 use RichId\AutoconfigureBundle\Factory\Partials\TagAnnotationServiceConfigurationFactory;
 use RichId\AutoconfigureBundle\Model\ServiceConfiguration;
@@ -25,6 +26,7 @@ final class ServiceConfigurationFactory
         ArgumentAnnotationServiceConfigurationFactory::class,
         DecorationAnnotationServiceConfigurationFactory::class,
         EventListenerServiceConfigurationFactory::class,
+        MethodAnnotationServiceConfigurationFactory::class,
         PropertyAnnotationServiceConfigurationFactory::class,
         TagAnnotationServiceConfigurationFactory::class,
     ];