+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 1.1.0
+
+- Add service tag injection
+- Add method annotation to add MethodCall to the service definition
+
 ## Version 1.0.1
 
 - Fix annotation dependency where it was